{"schema":"decodex/commit/1","summary":"split state account autonomy and git credential modules","authority":"manual"}
